requirements.txt
streamlit
tensorflow-cpu==2.12.0
opencv-python-headless
pillow
numpy